added: set_transform (etc.), transform, aggregate, bounding_box_query, transformations; set default value of mainintain_positioning to True
