+++ src/spatialdata/transformations/__init__.py
@@ -1,0 +1,29 @@
+from spatialdata.transformations.operations import (
+    get_transformation,
+    get_transformation_between_coordinate_systems,
+    remove_transformation,
+    set_transformation,
+)
+from spatialdata.transformations.transformations import (
+    Affine,
+    BaseTransformation,
+    Identity,
+    MapAxis,
+    Scale,
+    Sequence,
+    Translation,
+)
+
+__all__ = [
+    "BaseTransformation",
+    "Identity",
+    "MapAxis",
+    "Translation",
+    "Scale",
+    "Affine",
+    "Sequence",
+    "get_transformation",
+    "set_transformation",
+    "remove_transformation",
+    "get_transformation_between_coordinate_systems",
+]
