+++ tests/_core/test_models.py
@@ -23,11 +23,6 @@ from spatial_image import SpatialImage, to_spatial_image
 from xarray import DataArray
 
 from spatialdata import SpatialData
-from spatialdata._core.spatialdata_operations import (
-    get_transformation,
-    set_transformation,
-)
-from spatialdata._core.transformations import Scale
 from spatialdata._types import ArrayLike
 from spatialdata.models import (
     Image2DModel,
@@ -44,6 +39,11 @@ from spatialdata.transformations._utils import (
     _set_transformations,
     _set_transformations_xarray,
 )
+from spatialdata.transformations.operations import (
+    get_transformation,
+    set_transformation,
+)
+from spatialdata.transformations.transformations import Scale
 from tests._core.conftest import MULTIPOLYGON_PATH, POINT_PATH, POLYGON_PATH
 from tests.conftest import (
     _get_images,
