+++ tests/_core/test_rasterize.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from spatial_image import SpatialImage
 
-from spatialdata._core._rasterize import rasterize
+from spatialdata._core.operations.rasterize import rasterize
 from spatialdata._io._utils import _iter_multiscale
 from spatialdata.models import get_axis_names
 from spatialdata.models._utils import get_spatial_axes
