+++ tests/_core/test_transformations.py
@@ -5,17 +5,6 @@ import pytest
 import xarray.testing
 from xarray import DataArray
 
-from spatialdata._core.transformations import (
-    Affine,
-    BaseTransformation,
-    Identity,
-    MapAxis,
-    Scale,
-    Sequence,
-    Translation,
-    _decompose_affine_into_linear_and_translation,
-    _get_affine_for_element,
-)
 from spatialdata.models import Image2DModel, PointsModel
 from spatialdata.models._utils import ValidAxis_t
 from spatialdata.transformations.ngff._utils import get_default_coordinate_system
@@ -29,6 +18,17 @@ from spatialdata.transformations.ngff.ngff_transformations import (
     NgffScale,
     NgffSequence,
     NgffTranslation,
+)
+from spatialdata.transformations.transformations import (
+    Affine,
+    BaseTransformation,
+    Identity,
+    MapAxis,
+    Scale,
+    Sequence,
+    Translation,
+    _decompose_affine_into_linear_and_translation,
+    _get_affine_for_element,
 )
