+++ tests/_io/test_readwrite.py
@@ -12,12 +12,12 @@ from shapely.geometry import Point
 from spatial_image import SpatialImage
 
 from spatialdata import SpatialData
-from spatialdata._core.spatialdata_operations import (
+from spatialdata._io._utils import _are_directories_identical
+from spatialdata.transformations.operations import (
     get_transformation,
     set_transformation,
 )
-from spatialdata._core.transformations import Identity, Scale
-from spatialdata._io._utils import _are_directories_identical
+from spatialdata.transformations.transformations import Identity, Scale
 from tests.conftest import _get_images, _get_labels, _get_points, _get_shapes
